# reduce allocs for lookups of Percentile types