--- a/build.gradle
+++ b/build.gradle
@@ -100,7 +100,7 @@ subprojects {
     profilers = ['stack', 'gc']
     includeTests = false
     duplicateClassesStrategy = 'warn'
-    include = ['.*PatternMatching.*']
+    include = ['.*PercentileTimers.*']
   }
 
   checkstyle {