--- a/spectator-api/src/main/java/com/netflix/spectator/api/CompositeRegistry.java
+++ b/spectator-api/src/main/java/com/netflix/spectator/api/CompositeRegistry.java
@@ -105,6 +105,7 @@ public final class CompositeRegistry implements Registry {
       distSummaries.clear();
       timers.clear();
       gauges.clear();
+      state.clear();
     } finally {
       wlock.unlock();
     }